Fix wheel turn left. The value to send is 0x1FFFFF (21-bit), therefore the interfaces are refactored accordingly.

+++ lib/Rego6xx/Rego6xxStdRsp.h
@@ -96,7 +96,7 @@ public:
      * 
      * @return value
      */
-    uint16_t getValue() const;
+    uint32_t getValue() const;
 
 private:
